pyproject.toml
[project]
name = "pivotal-wildlife"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chardet>=5.2.0",
    "gradio>=5.9.1",
    "numpy>=2.2.1",
    "pillow>=11.0.0",
    "supervision>=0.25.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.23",
    "ultralytics-yolov5>=0.1.1",
    "wget>=3.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.391",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]
